+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2025-10-17
+### Added
+- Added delta scanning for pushes and pull requests
+- Added input scanMode to toggle delta/full scan
+### Changed
+- Upgraded scanoss-py version to v1.37.0 
+
 ## [1.2.5] - 2025-10-02
 ### Fixed
 - Missing brackets when initialising scanoss.json with link
@@ -143,3 +150,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [1.2.3]: https://github.com/scanoss/gha-code-scan/compare/v1.2.2...v1.2.3
 [1.2.4]: https://github.com/scanoss/gha-code-scan/compare/v1.2.3...v1.2.4
 [1.2.5]: https://github.com/scanoss/gha-code-scan/compare/v1.2.4...v1.2.5
+[1.3.0]: https://github.com/scanoss/gha-code-scan/compare/v1.2.5...v1.3.0